fixed delete and added like functionality

+++ src/components/Chirp.js
@@ -1,11 +1,13 @@
 import React, { useState } from 'react'
 import { FaTrash } from 'react-icons/fa'
-import { deleteChirp } from '../services/api-helper'
+import { deleteChirp, getChirpById, updateChirp } from '../services/api-helper'
 import Comment from './Comment'
 
 function Chirp(props) {
 
     const [deleted, setDeleted] = useState(false)
+    const [liked, setLiked] = useState(false)
+    const [numLikes, setNumLikes] = useState(props.numLikes)
 
     const comments = props.comments.map((comment, i) => {
         return (
@@ -15,11 +17,21 @@ function Chirp(props) {
         )
     })
 
-    const handleDeleteChirp = async (id) => {
-        console.log('this is getting clicked', props.id)
-        const json = await deleteChirp(props.id)
-        console.log('when clicked we get ', json)
-        if (json.status === 'success') {
+    const handleLike = async () => {
+        console.log('Chirp - handleLike - liked', liked)
+        if (!liked) {
+            const chirp = await getChirpById(props.id)
+            chirp.numLikes++
+            setNumLikes(chirp.numLikes)
+            updateChirp(props.id, chirp)
+            setLiked(true)
+        }
+    }
+
+    const handleDelete = () => {
+        console.log('Chirp - handleDelete - deleted', deleted)
+        if (!deleted) {
+            deleteChirp(props.id)
             setDeleted(true)
         }
     }
@@ -29,11 +41,16 @@ function Chirp(props) {
             {!deleted && <>
                 <p id="name">{props.username}</p>
                 <p>{props.body}</p>
-                <p>Date:{props.date}</p>
-                <p><i class="fas fa-hand-spock"></i> {props.numLikes}</p>
-                <FaTrash
-                    onClick={() => handleDeleteChirp(props._id)}
-                />
+                <p>Date:{props.date[0]}</p>
+                <p>Time:{props.date[1]}</p>
+                <p>
+                    <button onClick={handleLike}>
+                        <i class="fas fa-hand-spock"></i>
+                    </button> {numLikes}
+                </p>
+                <button><FaTrash
+                    onClick={handleDelete}
+                /></button>
                 <hr />
                 <ul>
                     {comments}
